+++ packages/renderer/threejs/src/index.ts
@@ -1,7 +1,8 @@
 export * from "./ModularSpaceRenderer";
 export * from "./RendererStateAdapter";
+export * from "./RenderPipeline";
+export * from "./types";
 
 export * from "./events";
 
 export * from "./utils/coordinateUtils";
-export * from "./utils/debounce";